FIX subprocess call for Windows
The call made with subprocess.POpen all used shell=False. This works on
Linux and mac, but is an issue on Windows. In this case Windows expects
a full string to the executable. The code only provides a shell command,
and we should set shell=True for Windows.

This has the drawback, that every item in the list is passed to the
shell and left to the shell to figure out. On Windows this is usually
alright as the cmd.exe is not very smart, but it could pose a problem
for other shells.